cloud_function/requirements.txt
Flask==2.2.5
PyJWT==2.8.0
cryptography
firebase-admin==6.3.0
